feat(schema-generator): add branded scalar types support
Implement branded scalar types for precise GraphQL scalar mapping:
- Add IDString, IDNumber, Int, Float branded types in @gqlkit-ts/runtime
- Add ScalarRegistry for type-to-scalar mapping management
- Add SymbolResolver for detecting @gqlkit-ts/runtime imports
- Add BrandedDetector for identifying branded scalar types
- Extend type-extractor to convert branded types to GraphQL scalars
- Extend resolver-extractor to support branded types in arguments
- Add comprehensive unit and integration tests

The number type maps to Float by default; use Int branded type for integers.

+++ packages/cli/src/resolver-extractor/extract-resolvers.ts
@@ -100,7 +100,7 @@ function collectDiagnostics(
 
 const PRIMITIVE_TYPE_MAP: Record<string, string> = {
   string: "String",
-  number: "Int",
+  number: "Float",
   boolean: "Boolean",
 };
